# Optimize hashing of leaf nodes