## Why this should be merged

TL;DR: Reduces the cost of hashing leaf nodes by ~20%

Before:
```
Benchmark_HashNode/empty_node-12         	12526160	        92.37 ns/op	       0 B/op	       0 allocs/op
Benchmark_HashNode/has_value-12          	12145953	        99.37 ns/op	       0 B/op	       0 allocs/op
Benchmark_HashNode/has_key-12            	13527973	        87.63 ns/op	       0 B/op	       0 allocs/op
Benchmark_HashNode/1_child-12            	 8485260	       141.7  ns/op	       0 B/op	       0 allocs/op
Benchmark_HashNode/2_children-12         	 5874465	       205.1  ns/op	       0 B/op	       0 allocs/op
Benchmark_HashNode/16_children-12        	 1216708	       979.0  ns/op	       0 B/op	       0 allocs/op

Benchmark_EncodeDBNode/empty_node-12         	27059378	        44.58 ns/op	       2 B/op	       1 allocs/op
Benchmark_EncodeDBNode/has_value-12          	19252648	        61.28 ns/op	       8 B/op	       1 allocs/op
Benchmark_EncodeDBNode/1_child-12            	 6188838	       191.8  ns/op	      48 B/op	       1 allocs/op
Benchmark_EncodeDBNode/2_children-12         	 4343116	       279.8  ns/op	      96 B/op	       1 allocs/op
Benchmark_EncodeDBNode/16_children-12        	  729697	      1700    ns/op	     640 B/op	       1 allocs/op
```

After:
```
Benchmark_HashNode/empty_node-12         	15640392	        75.73 ns/op	       0 B/op	       0 allocs/op
Benchmark_HashNode/has_value-12          	14471510	        83.53 ns/op	       0 B/op	       0 allocs/op
Benchmark_HashNode/has_key-12            	16408834	        72.83 ns/op	       0 B/op	       0 allocs/op
Benchmark_HashNode/1_child-12            	 8587516	       139.6  ns/op	       0 B/op	       0 allocs/op
Benchmark_HashNode/2_children-12         	 5899671	       201.5  ns/op	       0 B/op	       0 allocs/op
Benchmark_HashNode/16_children-12        	 1244270	       965.4  ns/op	       0 B/op	       0 allocs/op

Benchmark_EncodeDBNode/empty_node-12         	37150839	        32.81 ns/op	       2 B/op	       1 allocs/op
Benchmark_EncodeDBNode/has_value-12          	24834051	        49.13 ns/op	       8 B/op	       1 allocs/op
Benchmark_EncodeDBNode/1_child-12            	 6342018	       192.8  ns/op	      48 B/op	       1 allocs/op
Benchmark_EncodeDBNode/2_children-12         	 4295674	       279.0  ns/op	      96 B/op	       1 allocs/op
Benchmark_EncodeDBNode/16_children-12        	  746523	      1643    ns/op	     640 B/op	       1 allocs/op
```

## How this works

If the node doesn't have any children, we can skip allocating the `keys` slice (even on the stack).

## How this was tested

- [x] CI